clinker-core: extract pipeline, aggregate, and transform config modules
Complete the config/mod.rs decomposition by relocating the heaviest,
single-topic cluster, leaving mod.rs a thin module-declaration +
re-export hub:

- pipeline.rs  — PipelineConfig / PipelineMeta and the whole compile
                 pipeline that lowers a parsed config into an execution
                 plan (compile, source-config extraction, node lowering,
                 post-deserialization validation, and the error-handling /
                 DLQ config), plus MemoryConfig, MetricsConfig, and
                 ConcurrencyConfig
- aggregate.rs — Aggregate and Combine strategy hints, AggregateConfig,
                 and the correlation-key types
- transform.rs — Transform inputs/views, validation entries, and log
                 directives

After this change config/mod.rs only declares the submodules and
re-exports them, so a single-cluster edit rebuilds just the touched
submodule instead of the whole config translation unit.

Pure relocation: no symbol renames and no public-API change. Every
existing `clinker_core::config::*` path still resolves through the
re-exports.

Closes #257, #129
