+++ crates/clinker-core/src/config/transform.rs
@@ -1,0 +1,155 @@
+//! Transform node inputs, validation directives, and log directives.
+
+use indexmap::IndexMap;
+use serde::de;
+use serde::{Deserialize, Deserializer, Serialize};
+
+/// Input wiring for a transform — specifies which upstream transform(s) feed records.
+///
+/// String values become `Single`; arrays become `Multiple`.
+/// Custom deserialization handles both forms.
+#[derive(Debug, Clone, Serialize)]
+pub enum TransformInput {
+    /// Single upstream: `"categorize.high_value"` or `"transform_name"`.
+    Single(String),
+    /// Multiple upstreams (union): `["branch_a", "branch_b"]`.
+    Multiple(Vec<String>),
+}
+
+impl<'de> Deserialize<'de> for TransformInput {
+    fn deserialize<D: Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
+        struct TransformInputVisitor;
+
+        impl<'de> de::Visitor<'de> for TransformInputVisitor {
+            type Value = TransformInput;
+
+            fn expecting(&self, formatter: &mut std::fmt::Formatter) -> std::fmt::Result {
+                formatter.write_str("a string or array of strings")
+            }
+
+            fn visit_str<E: de::Error>(self, v: &str) -> Result<Self::Value, E> {
+                Ok(TransformInput::Single(v.to_owned()))
+            }
+
+            fn visit_seq<A: de::SeqAccess<'de>>(self, mut seq: A) -> Result<Self::Value, A::Error> {
+                let mut items = Vec::new();
+                while let Some(item) = seq.next_element::<String>()? {
+                    items.push(item);
+                }
+                if items.is_empty() {
+                    return Err(de::Error::custom(
+                        "transform input array must not be empty (use a single string for one upstream, or omit for default flow)",
+                    ));
+                }
+                Ok(TransformInput::Multiple(items))
+            }
+        }
+
+        deserializer.deserialize_any(TransformInputVisitor)
+    }
+}
+
+/// A declarative validation attached to a transform.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(deny_unknown_fields)]
+pub struct ValidationEntry {
+    pub name: Option<String>,
+    pub field: Option<String>,
+    pub check: String,
+    pub args: Option<IndexMap<String, serde_json::Value>>,
+    #[serde(default = "default_severity")]
+    pub severity: ValidationSeverity,
+    pub message: Option<String>,
+}
+
+fn default_severity() -> ValidationSeverity {
+    ValidationSeverity::Error
+}
+
+impl ValidationEntry {
+    /// Auto-derive name from field and check if not specified.
+    pub fn resolved_name(&self) -> String {
+        self.name.clone().unwrap_or_else(|| match &self.field {
+            Some(f) => format!("{}:{}", f, self.check),
+            None => self.check.clone(),
+        })
+    }
+}
+
+/// Validation severity: error routes to DLQ, warn logs and continues.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
+pub enum ValidationSeverity {
+    #[serde(rename = "error")]
+    Error,
+    #[serde(rename = "warn")]
+    Warn,
+}
+
+/// A logging directive attached to a transform.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(deny_unknown_fields)]
+pub struct LogDirective {
+    pub level: LogLevel,
+    pub when: LogTiming,
+    pub condition: Option<String>,
+    pub message: String,
+    pub fields: Option<Vec<String>>,
+    pub every: Option<u64>,
+    pub log_rule: Option<String>,
+}
+
+/// When a log directive fires.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
+pub enum LogTiming {
+    #[serde(rename = "before_transform")]
+    BeforeTransform,
+    #[serde(rename = "after_transform")]
+    AfterTransform,
+    #[serde(rename = "per_record")]
+    PerRecord,
+    #[serde(rename = "on_error")]
+    OnError,
+}
+
+/// Log level for directives (YAML config domain).
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
+pub enum LogLevel {
+    #[serde(rename = "trace")]
+    Trace,
+    #[serde(rename = "debug")]
+    Debug,
+    #[serde(rename = "info")]
+    Info,
+    #[serde(rename = "warn")]
+    Warn,
+    #[serde(rename = "error")]
+    Error,
+}
+
+/// Lightweight read-only view over a transform-like node
+/// (`Transform`, `Aggregate`, `Route`) yielded by
+/// [`PipelineConfig::transform_views`]. Carries the fields the Kiln IDE
+/// and schema-validation passes need; callers that need variant-specific
+/// bodies (`TransformBody`, `AggregateBody`, etc.) should walk
+/// [`PipelineConfig::nodes`] directly.
+#[derive(Debug, Clone, Copy)]
+pub struct TransformView<'a> {
+    pub name: &'a str,
+    pub description: Option<&'a str>,
+    pub cxl_source: &'a str,
+    pub notes: Option<&'a serde_json::Value>,
+    pub kind: TransformViewKind,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum TransformViewKind {
+    Transform,
+    Aggregate,
+    Route,
+}
+
+impl<'a> TransformView<'a> {
+    pub fn cxl_source(&self) -> &'a str {
+        self.cxl_source
+    }
+}
